+++ partnerusagereportsv1/partner_usage_reports_v1_suite_test.go
@@ -1,5 +1,5 @@
 /**
- * (C) Copyright IBM Corp. 2020.
+ * (C) Copyright IBM Corp. 2024.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,16 @@
  * limitations under the License.
  */
 
-package common
+package partnerusagereportsv1_test
 
-// Version of the SDK
-const Version = "0.59.0"
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestPartnerUsageReportsV1(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "PartnerUsageReportsV1 Suite")
+}